remade new front-end project on android studio with basic views
