+++ .idea/workspace.xml
@@ -4,7 +4,12 @@
     <option name="autoReloadType" value="NONE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="f465b897-2635-4abc-8be7-ea564e542ecb" name="Changes" comment="" />
+    <list default="true" id="f465b897-2635-4abc-8be7-ea564e542ecb" name="Changes" comment="">
+      <change beforePath="$PROJECT_DIR$/front-end/.idea/gradle.xml" beforeDir="false" afterPath="$PROJECT_DIR$/front-end/.idea/gradle.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/front-end/.idea/misc.xml" beforeDir="false" afterPath="$PROJECT_DIR$/front-end/.idea/misc.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/front-end/app/build.gradle" beforeDir="false" afterPath="$PROJECT_DIR$/front-end/app/build.gradle" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/front-end/gradle.properties" beforeDir="false" afterPath="$PROJECT_DIR$/front-end/gradle.properties" afterDir="false" />
+    </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
@@ -26,7 +31,8 @@
     "RunOnceActivity.OpenProjectViewOnStart": "true",
     "RunOnceActivity.ShowReadmeOnStart": "true",
     "RunOnceActivity.cidr.known.project.marker": "true",
-    "cidr.known.project.marker": "true"
+    "cidr.known.project.marker": "true",
+    "last_opened_file_path": "/Users/louietang/AndroidStudioProjects/404-group-not-found"
   }
 }]]></component>
   <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
