+++ grocerymanageras/app/src/main/AndroidManifest.xml
@@ -10,11 +10,13 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Grocerymanager"
+        android:theme="@style/Theme.Grocerymanageras"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
-            android:exported="true">
+            android:exported="true"
+            android:label="@string/app_name"
+            android:theme="@style/Theme.Grocerymanageras">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
