+++ grocerymanageras/app/src/main/java/com/example/grocerymanageras/FirstFragment.java
@@ -1,4 +1,4 @@
-package com.example.grocerymanager;
+package com.example.grocerymanageras;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -9,7 +9,7 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.navigation.fragment.NavHostFragment;
 
-import com.example.grocerymanager.databinding.FragmentFirstBinding;
+import com.example.grocerymanageras.databinding.FragmentFirstBinding;
 
 public class FirstFragment extends Fragment {
