+++ grocerymanageras/app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="com.example.grocerymanager.FirstFragment"
+        android:name="com.example.grocerymanageras.FirstFragment"
         android:label="@string/first_fragment_label"
         tools:layout="@layout/fragment_first">
 
@@ -17,7 +17,7 @@
     </fragment>
     <fragment
         android:id="@+id/SecondFragment"
-        android:name="com.example.grocerymanager.SecondFragment"
+        android:name="com.example.grocerymanageras.SecondFragment"
         android:label="@string/second_fragment_label"
         tools:layout="@layout/fragment_second">
