+++ grocerymanageras/app/src/main/res/values-land/dimens.xml
