+++ grocerymanageras/app/src/main/res/values-v23/themes.xml
@@ -1,6 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
 
-    <style name="Theme.Grocerymanager" parent="Base.Theme.Grocerymanager">
+    <style name="Theme.Grocerymanageras" parent="Base.Theme.Grocerymanageras">
         <!-- Transparent system bars for edge-to-edge. -->
         <item name="android:navigationBarColor">@android:color/transparent</item>
         <item name="android:statusBarColor">@android:color/transparent</item>
