+++ grocerymanageras/app/src/main/res/values-w1240dp/dimens.xml
