+++ grocerymanageras/app/src/main/res/values-w600dp/dimens.xml
